chroe : shared index 분리
