+++ src/shared/ui/Card.tsx
@@ -1,0 +1,6 @@
+import { forwardRefWithType } from "../../types"
+
+export const Card = forwardRefWithType<"div">(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+Card.displayName = "Card"
