+++ src/shared/ui/SelectItem.tsx
@@ -1,0 +1,25 @@
+import { Check } from "lucide-react"
+import * as SelectPrimitive from "@radix-ui/react-select"
+import { forwardRef } from "react"
+
+// SelectPrimitive.Item의 props 타입과 ref 타입을 추출
+type SelectItemProps = React.ComponentProps<typeof SelectPrimitive.Item>
+
+export const SelectItem = forwardRef<React.ElementRef<typeof SelectPrimitive.Item>, SelectItemProps>(
+  ({ className, children, ...props }, ref) => (
+    <SelectPrimitive.Item
+      ref={ref}
+      className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
+      {...props}
+    >
+      <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
+        <SelectPrimitive.ItemIndicator>
+          <Check className="h-4 w-4" />
+        </SelectPrimitive.ItemIndicator>
+      </span>
+      <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
+    </SelectPrimitive.Item>
+  ),
+)
+
+SelectItem.displayName = SelectPrimitive.Item.displayName
