+++ src/shared/ui/TableBody.tsx
@@ -1,0 +1,5 @@
+import { forwardRefWithType } from "../../types"
+
+export const TableBody = forwardRefWithType<"tbody">(({ className, ...props }, ref) => (
+  <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+))
