+++ src/shared/ui/TableHead.tsx
@@ -1,0 +1,9 @@
+import { forwardRefWithType } from "../../types"
+
+export const TableHead = forwardRefWithType<"th">(({ className, ...props }, ref) => (
+  <th
+    ref={ref}
+    className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
+    {...props}
+  />
+))
