+++ src/shared/ui/TableHeader.tsx
@@ -1,0 +1,5 @@
+import { forwardRefWithType } from "../../types";
+
+export const TableHeader = forwardRefWithType<"thead">(({ className, ...props }, ref) => (
+    <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />
+  ))
